backend-flask/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.10.1
supabase==2.0.2
flask-limiter==3.5.0
Werkzeug==2.3.7
email-validator==2.0.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
